New release build scripts (used for 5.3.0-M1/M2):
The ax-release.py script file uses a release plan, which is just a
config file mapping repository names to the to-be-released version
numbers. From the version number the branch is automatically derived.

There is also an ax-set-snapshot-versions.py script, to reset the
revision and dependencies back to snapshots. This is based on the
branches that you have currently checked out.